Merged PR 973: Raft tutorial using Azure Service Bus
This PR refactors this repository to make it very focused on a small number of samples (similar to how [dapr](https://github.com/dapr/samples) is doing it). This PR also introduces tutorials number 4, 5 and 6. The actual tutorial is not yet done, I have only written the samples (but added lots of documentation in them).

These 3 samples are building on top of each other, and reuse code.