Merged PR 1011: The HelloWorldTasks sample
This is the code for the **_Hello World_** sample with the **_AsyncTasks_ _Programming Model._**

I adapted a very simple existing example to store asynchronously strings in a shared object. The expected result is that the shared store must contain the string "Hello World!". However, due to a bug in the code, in some executions of this program the result is a different string.

The `Specification.Assert()` call makes sure that any bug of this kind will not slip unnoticed. Indeed, once in a while when running

  `dotnet HelloWorldTasks.dll` (for netcoreapp2.2)

or

  `HelloWorldTasks.exe`  (for .Net 4.7 or 4.6)

we get this:

    Microsoft.Coyote.Runtime.AssertionFailureException: Value is 'Good Morning' instead of 'Hello World!'.