Merged PR 989: Replace ICommunicationManager with a ClusterManager state machine
Replace ICommunicationManager with a ClusterManager state machine, moving most of the Azure implementation to "Raft.Azure" in a subclass of ClusterManager called AzureClusterManager.  This simplifies ServerHost such that it now only deals with creating the Message Bus, subscription, creating Server machine and starting the ReceiveMessageAsync loop. Similarly ServerHost is moved to Raft.AzureClient, renamed to AzureServer and so Raft is now Azure neutral.

Also removed Program.cs from Raft and changed Raft into a ClassLibrary, I did this by merging client and server functionality into the Raft.Azure sample, it decides which mode it is in based on additional command line arguments.  The Raft.Mocking library never needed Raft to be executable.

Also made it easier to setup the sample by providing a powershell script that sets up a new resource group and service bus namespace and programatically create the needed Topic.