Merged PR 1358: fixes in cloud messaging sample script
This PR fixes the cloud messaging sample script by changing the testing strategy to random and taming the timer using `--timeout-delay 10` (this is an advanced testing option that the higher it is the lower the probability the timer will timeout during testing). We can switch to RL once we have that.

Also some cleanup: removed unused usings and removed a tutorial that does not exist from the list.